RPCT-4140: Added ststic_cast

+++ libs/TrendView/TrendParam.cpp
@@ -340,7 +340,7 @@ namespace TrendLib
 
 	void TrendParam::setLaneDuration(qint64 value)
 	{
-		m_duration = std::clamp(value, 500_ms, 24_hours * 365);
+		m_duration = std::clamp(value, std::static_cast<qint64>(500_ms), std::static_cast<qint64>(24_hours * 365));
 	}
 
 	int TrendParam::hightlightRulerIndex() const
